+++ app/dashboard/page.tsx
@@ -3,7 +3,7 @@
 import { HiArrowSmRight, HiChartPie, HiInbox, HiShoppingBag, HiTable, HiUser, HiViewBoards,HiQuestionMarkCircle } from "react-icons/hi";
 
 import { Sidebar } from "flowbite-react";
-import { DashboardIcon } from "./svg.jsx";
+import { AddDocIcon, DashboardIcon } from "./svg";
 
 
 export default function Dashboard() {
@@ -161,17 +161,25 @@ export function SideSet() {
             </p>
          </div>
       </div>
-      <div className="flex items-center justify-center h-48 mb-4 rounded bg-gray-50 dark:bg-gray-800">
-         <p className="text-2xl text-gray-400 dark:text-gray-500">
-            <svg className="w-3.5 h-3.5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 18 18">
-               <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 1v16M1 9h16"/>
-            </svg>
-         </p>
-      </div>
+
 
    </div>
+   <AddDocumentButton/>
 </div>
 
 </div>
   );
+}
+
+function AddDocumentButton() {
+
+  return(
+    <button className="bg-sky-600 p-4 fixed top-3/4 left-3/4 rounded-full shadow-md transition ease-in-out duration-200 hover:bg-lime-700" title="add document">
+      <div>
+        <AddDocIcon/>
+
+      </div>
+    </button>
+  )
+
 }
