--- app/dashboard/svg.jsx
@@ -1,23 +1,0 @@
-export function DashboardIcon() {
-
-
-  return (
-    <svg
-      class="h-4 w-4 text-slate-500"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      stroke-width="2"
-      stroke="currentColor"
-      fill="none"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    >
-      {" "}
-      <path stroke="none" d="M0 0h24v24H0z" />{" "}
-      <rect x="4" y="4" width="16" height="16" rx="2" />{" "}
-      <line x1="4" y1="10" x2="20" y2="10" />{" "}
-      <line x1="10" y1="4" x2="10" y2="20" />
-    </svg>
-  );
-}
