+++ app/dashboard/svg.tsx
@@ -1,0 +1,47 @@
+export function DashboardIcon() {
+
+
+  return (
+    <svg
+      className="h-4 w-4 text-slate-50"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      stroke-width="2"
+      stroke="currentColor"
+      fill="none"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    >
+      {" "}
+      <path stroke="none" d="M0 0h24v24H0z" />{" "}
+      <rect x="4" y="4" width="16" height="16" rx="2" />{" "}
+      <line x1="4" y1="10" x2="20" y2="10" />{" "}
+      <line x1="10" y1="4" x2="10" y2="20" />
+    </svg>
+  );
+}
+
+
+export function AddDocIcon() {
+
+  return (
+    <svg
+      className="h-8 w-8 text-gray-50 "
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      stroke-width="2"
+      stroke="currentColor"
+      fill="none"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    >
+      {" "}
+      <path stroke="none" d="M0 0h24v24H0z" />{" "}
+      <line x1="12" y1="5" x2="12" y2="19" />{" "}
+      <line x1="5" y1="12" x2="19" y2="12" />
+    </svg>
+  );
+
+}
